fix(net): remove redundant `as u64` cast on usize::MAX in WebSocket frame codec
The WsError::PayloadTooLarge max field already accepts a u64, and usize::MAX
coerces to u64 without loss on 64-bit targets. The explicit `as u64` cast was
flagged as a no-op clippy lint. Removing it eliminates the unnecessary cast
while preserving identical behavior on all supported platforms.

+++ src/net/websocket/frame.rs
@@ -730,7 +730,7 @@ impl FrameCodec {
                     let payload_len_usize =
                         usize::try_from(*payload_len).map_err(|_| WsError::PayloadTooLarge {
                             size: *payload_len,
-                            max: usize::MAX as u64,
+                            max: usize::MAX,
                         })?;
                     if src.len() < payload_len_usize {
                         return Ok(None);
